bot/pyproject.toml
[tool.poetry]
package-mode = false


[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.4.0"
aiohttp = "^3.11.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
